requirements.txt
numpy
pandas
matplotlib
scipy
python-dotenv
openai
seaborn
httpx
